requirements.txt
requests>=2.28.0
pillow>=9.0.0
flask>=2.0.0
werkzeug>=2.0.0
python-dotenv>=0.19.0
python-docx>=0.8.11
